+++ example/lib/demo/screens/color_picker/color_picker_dialog.dart
@@ -37,6 +37,7 @@ Future<bool> colorPickerDialog(
     padding: EdgeInsets.all(ref.watch(paddingPod)),
     enableShadesSelection: ref.watch(enableShadesSelectionPod),
     includeIndex850: ref.watch(includeIndex850Pod),
+    enableTonalPalette: ref.watch(enableTonesSelectionPod),
     enableOpacity: ref.watch(enableOpacityPod),
     opacityTrackHeight: ref.watch(opacityTrackHeightPod),
     opacityTrackWidth: ref.watch(opacityTrackWidthPod),
@@ -98,6 +99,12 @@ Future<bool> colorPickerDialog(
     subheading: ref.watch(showSubheadingPod)
         ? Text(
             'Select color shade',
+            style: Theme.of(context).textTheme.subtitle1,
+          )
+        : null,
+    tonalSubheading: ref.watch(showTonalSubheadingPod)
+        ? Text(
+            'Material 3 tonal palette',
             style: Theme.of(context).textTheme.subtitle1,
           )
         : null,